requirements/local.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.0
billiard==3.3.0.23
boto==2.43.0
celery==3.1.23
coreapi==2.1.1
Django==1.8
django-celery==3.1.17
django-ckeditor==5.1.1
django-cors-headers==1.1.0
#--- ONLY FOR LOCAL
django-debug-panel==0.8.3
#--- ONLY FOR LOCAL
django-debug-toolbar==1.4
django-extensions==1.6.7
django-ses==0.8.1
django-smtp-ssl==1.0
django-user-agents==0.3.0
djangorestframework==3.5.3
django-rest-swagger==2.1.1
#--- ONLY FOR LOCAL
Fabric==1.12.0
gunicorn==19.4.5
#--- ONLY FOR LOCAL
ipdb==0.10.1
pathlib2==2.1.0
Pillow==3.2.0
#--- ONLY FOR LOCAL
pygraphviz==1.3.1
ptyprocess==0.5.1
pytz==2016.4
redis==2.10.5
requests==2.10.0
simplejson==3.10.0
six==1.10.0
sqlparse==0.1.19
#--- ONLY FOR LOCAL
Sphinx==1.5.3
supervisor==3.3.0
ua-parser==0.7.1
uritemplate==3.0.0
user-agents==1.0.1